refactor: make `cookie` into Context/Request
