+++ package.json
@@ -22,16 +22,12 @@
     "./basic-auth": "./dist/middleware/basic-auth/index.js",
     "./bearer-auth": "./dist/middleware/bearer-auth/index.js",
     "./body-parse": "./dist/middleware/body-parse/index.js",
-    "./cookie": "./dist/middleware/cookie/index.js",
     "./cors": "./dist/middleware/cors/index.js",
     "./etag": "./dist/middleware/etag/index.js",
-    "./graphql-server": "./dist/middleware/graphql-server/index.js",
     "./html": "./dist/middleware/html/index.js",
     "./jsx": "./dist/middleware/jsx/index.js",
     "./jwt": "./dist/middleware/jwt/index.js",
     "./logger": "./dist/middleware/logger/index.js",
-    "./mustache": "./dist/middleware/mustache/index.js",
-    "./mustache.module": "./dist/middleware/mustache/module.mjs",
     "./powered-by": "./dist/middleware/powered-by/index.js",
     "./pretty-json": "./dist/middleware/pretty-json/index.js",
     "./serve-static": "./dist/middleware/serve-static/index.js",
