+++ src/middleware.ts
@@ -2,7 +2,6 @@
 export * from './middleware/basic-auth'
 export * from './middleware/bearer-auth'
 export * from './middleware/body-parse'
-export * from './middleware/cookie'
 export * from './middleware/cors'
 export * from './middleware/etag'
 export * from './middleware/html'
