+++ src/request.ts
@@ -1,3 +1,6 @@
+import type { Cookie } from './utils/cookie'
+import { parse } from './utils/cookie'
+
 declare global {
   interface Request<ParamKeyType extends string = string> {
     param: {
@@ -16,6 +19,10 @@ declare global {
     header: {
       (name: string): string
       (): Record<string, string>
+    }
+    cookie: {
+      (name: string): string
+      (): Cookie
     }
   }
 }
@@ -74,4 +81,15 @@ export function extendRequestPrototype() {
       return result
     }
   } as InstanceType<typeof Request>['queries']
+
+  Request.prototype.cookie = function (this: Request, key?: string) {
+    const cookie = this.headers.get('Cookie') || ''
+    const obj = parse(cookie)
+    if (key) {
+      const value = obj[key]
+      return value
+    } else {
+      return obj
+    }
+  } as InstanceType<typeof Request>['cookie']
 }
