+++ src/utils/cookie.ts
@@ -1,21 +1,3 @@
-import type { Context } from '../../context'
-import type { Next } from '../../hono'
-
-declare global {
-  interface Request {
-    cookie: {
-      (name: string): string
-      (): Record<string, string>
-    }
-  }
-}
-
-declare module '../../context' {
-  interface Context {
-    cookie: (name: string, value: string, options?: CookieOptions) => void
-  }
-}
-
 export type Cookie = Record<string, string>
 export type CookieOptions = {
   domain?: string
@@ -28,27 +10,7 @@ export type CookieOptions = {
   sameSite?: 'Strict' | 'Lax' | 'None'
 }
 
-export const cookie = () => {
-  return async (c: Context, next: Next) => {
-    c.req.cookie = ((name?: string): string | Record<string, string> => {
-      const cookie = c.req.headers.get('Cookie') || ''
-      const obj = parse(cookie)
-      if (name) {
-        const value = obj[name]
-        return value
-      } else {
-        return obj
-      }
-    }) as typeof c.req.cookie
-    c.cookie = (name: string, value: string, opt?: CookieOptions) => {
-      const cookie = serialize(name, value, opt)
-      c.header('Set-Cookie', cookie)
-    }
-    await next()
-  }
-}
-
-const parse = (cookie: string): Cookie => {
+export const parse = (cookie: string): Cookie => {
   const pairs = cookie.split(/;\s*/g)
   const parsedCookie: Cookie = {}
   for (let i = 0, len = pairs.length; i < len; i++) {
@@ -58,7 +20,7 @@ const parse = (cookie: string): Cookie => {
   return parsedCookie
 }
 
-const serialize = (name: string, value: string, opt: CookieOptions = {}): string => {
+export const serialize = (name: string, value: string, opt: CookieOptions = {}): string => {
   value = encodeURIComponent(value)
   let cookie = `${name}=${value}`
